requirements.txt
pandas_datareader
pandas
numpy
pytrends
python-dateutil
datetime
sklearn